RI-6049 separate timezone update from dateFormat update
