+++ redisinsight/ui/src/pages/settings/components/general-settings/datetime-formatter/components/timezone-form/styles.module.scss
@@ -1,0 +1,3 @@
+.datetimeInput {
+  width: 240px;
+}
